Exchange object refactoring, some additional error handling
