+++ app/uk/gov/hmrc/fsetlaunchpadgateway/connectors/launchpad/exchangeobjects/interview/SeamlessLoginLink.scala
@@ -1,0 +1,10 @@
+package uk.gov.hmrc.fsetlaunchpadgateway.connectors.launchpad.exchangeobjects.interview
+
+import play.api.libs.json.Json
+
+case class SeamlessLoginLink(url: String, status: String, message: String)
+
+object SeamlessLoginLink {
+  implicit val seamlessLoginLinkFormat = Json.format[SeamlessLoginLink]
+}
+
